몬스터 시간에 따른 스탯 증가 보완
실시간으로 모든 몬스터에게 적용 됨,
미사일 콜라이더 trigger 끔

+++ Assets/0.Scripts/MonsterScripts/Monster.cs
@@ -40,13 +40,12 @@ public abstract class Monster : MonoBehaviour
     [Header("Player Item Effect")]
     public GameObject explosionEffect;
 
-
     public virtual void Init()
     {
         GameMgr gameMgr = GameObject.Find("GameMgr").GetComponent<GameMgr>();
         monMaxHP = monBasicMaxHP + ((monBasicMaxHP * 0.3f) * (int)(gameMgr.playTime / 60)); //시간별 몬스터 최대 체력 조절
         monCurHP = monMaxHP;
-        attackPower = basicAttackPower + ((basicAttackPower * 0.2f) * (gameMgr.playTime / 60));
+        attackPower = basicAttackPower + ((basicAttackPower * 0.2f) * (int)(gameMgr.playTime / 60));
         monStat = MonsterStat.Spawn;
     }
